Added support for the `iterator' interface for storages. Specifically,
_loadSerialEx(): New method which is similar to the API method
loadSerial() except that it returns both the pickle and the version
string for the specified oid+serial pair.  This is not a public
method.

loadSerial(): Reimplement in terms of _loadSerialEx() to reduce code
duplication.

iterator(): Public method to return a "transactions iterator".

_nexttxn(): Helper method to return the transaction metadata for the
transaction following the given in the argument.  Raises IndexError if
there is no next transaction.

_alltxnoids(): Helper method to return a list of all the oids modified
in the given transaction.  Note that the Full Berkeley storage has the
implied semantics that if an object is modified more than once in a
transaction, only the last such modification is retained.

Classes _TransactionsIterator, _RecordsIterator, _Record are private
helper classes to support the iterator interface.


git-svn-id: svn://svn.zope.org/repos/main@20550 62d5b8a3-27da-0310-9561-8e5933582275