Changed product import and install code.
Now, when products are installed, they are installed in alphabetical order by product name, regardless of which Products directory (e.g. INSTANCE_HOME or SOFTWARE_HOME) that they're in EXCEPT FOR PluginIndexes, which are always first, due to the requirement that they register interfaces with the product context machinery before ZCatalog is initialized.


git-svn-id: svn://svn.zope.org/repos/main@20590 62d5b8a3-27da-0310-9561-8e5933582275