DocumentClass was changed to support alternate locates a while back.  One of the ways in which this was done was to use string.punctuation with locale support turned on, stuffing the results into regular expressions.
Unfortunately, the insertion of the punctuation into regular expressions was completely literal.  This meant that, depending on locale, a regex could take on a completely different semantic due to the fact that it could include "]", "?", etc.  At times, this could cause a segmentation fault if a nonsensical generated regex was especially juicy.

I took out the code which localized punctuation, although the code which localizes letters is still there.


git-svn-id: svn://svn.zope.org/repos/main@20632 62d5b8a3-27da-0310-9561-8e5933582275