DocumentClass was changed to support alternate locates a while back.  One of the  ways in which this was done was to use string.punctuation with locale support t urned on, stuffing the results into regular expressions.
Unfortunately, the insertion of the punctuation into regular expressions was com
pletely literal.  This meant that, depending on locale, a regex could take on a
completely different semantic due to the fact that it could include "]", "?", et
c.  At times, this could cause a segmentation fault if a nonsensical generated r
egex was especially juicy.

I took out the code which localized punctuation, although the code which localiz
es letters is still there.


git-svn-id: svn://svn.zope.org/repos/main@20633 62d5b8a3-27da-0310-9561-8e5933582275