Prevent treating WikiNames that happen to resolve to Zope infrastructure methods - like ZopeTime - from breaking a page.
Before this fix, just putting the word ZopeTime on a page, unquoted,
would cause the page to raise an error.

In addition, don't handle this recognizing the condition and enabling
people to create a page with that name.  Otherwise, those pages will
shadow the method, and could wreak havoc.  (In the 'ZopeTime' case,
it's used in Wiki code and creation as a page would lead to an
infinite recursion in any page in the wiki or below.)

._simplewikilink_replace(), ._wikilink_replace(): If the object
matching the name is acquired, and doesn't have an 'absolute_url'
attribute, return the name marked with a red 'x' superscript.  It's a
bit bogus, but the editor can always '!' escape the reference.

This needs to be documented!


git-svn-id: svn://svn.zope.org/repos/main@33859 62d5b8a3-27da-0310-9561-8e5933582275