A few checkins, but some sans notices (the checkin notifications is implemented by the code i'm changing, sometimes breaking:), all aiming for:
Put the user's full name (when available) in the From: header.  To do
so we had to switch over to using sendmail, futz with getting the
subject line at the top of the file, etc.  Sigh.

There's a problem on the server - sendmail's '-F' option isn't having
effect.  Once it does, this should work...


git-svn-id: svn://svn.zope.org/repos/main@31498 62d5b8a3-27da-0310-9561-8e5933582275