Changing the catchall to digicool-cvs@zope.org.
Note to list members - checkin notices coming to this list should be
an exceptional thing, to be rectified once it's noticed.  Every
repository area should be fit within the auspices of some specific
project, with a mailling list for its checkin notices.  It's only the
ones that mistakenly weren't provided for that would wind up here -
and once noticed, they're supposed to be rectified.

If anyone feels this is an abuse of this list, let me know and i'll
create a separate list.  I'd rather *not* do anything to discourage
people from following this list (and providing some degree of "many
eyes" safety)...

Reply to my email address, not the baymountain.com address (we'll be
fixing that, hopefully within the next couple of days).

Ken
klm@digicool.com


git-svn-id: svn://svn.zope.org/repos/main@31509 62d5b8a3-27da-0310-9561-8e5933582275