Provide for special actions associated with a checkin match - a new traffic_table entry field "specials".  I'll be using this to effect processing of a repository symlinks table, which i'll be checking in shortly.
In the process of doing this, i factored out a routine for sending
email messages, for use in reporting the results of the special
processing and also for use in the already existing mail-notifications
stuff.

(Going over the code, there's a lot of grossness.  Much of the
functionality is good, it would be nice to step back and clean it up,
sometime, sigh...)


git-svn-id: svn://svn.zope.org/repos/main@31516 62d5b8a3-27da-0310-9561-8e5933582275