Use the new adjustlinks.LinkManager functionality to identify all effective containers for a path, so checkin notices can be sent to all parties concerned with a file, not just those that happen to subscribe to the hierarchy where a shared file was checked-in.
Eg, if someone checks-in something within Zope/lib/python/ZODB, then
notices will be sent to people who subscribe to Zope checkins and to
StandaloneZODB checkins - since they both link to Packages/ZODB, where
the stuff really lives.  This also happens if the checkin was done in
StandaloneZODB, or directly to Packages/ZODB...  This means the
interdependencies are taken care of by the system, according to the
replinks table.  The traffic_table gets really simple - you just
identify the top-level items, and changes to any of the contents are
properly attributed.

Also, stripped out much of the rsync machinery.  There may be more
residue of that, and still other cleanup to do.

Also, since we're now depending python 2.x, i'm assuming 're' is
available, and removing regex fallback provisions.


git-svn-id: svn://svn.zope.org/repos/main@31529 62d5b8a3-27da-0310-9561-8e5933582275