Another tiny fix, to circumvent sendmail error warnings that were cluttering our run log.
Switch sendmail command over to _not_ use "arpanet" mode, but still
collect the From line from the message.  Setting file permissions to
do the arpanet mode operation without sendmail error messages is a
pain, and we can use a combination of command line args and specifying
the right stuff in the file to get it to work (thanks to baymountain
guys for finding a combo that threaded the twisty maze).  PITA to find
the right combination that accommodates sendmail, but it works.


git-svn-id: svn://svn.zope.org/repos/main@31543 62d5b8a3-27da-0310-9561-8e5933582275