Be more specific about exception begin caugh in test_merge().
The only exceptions that an _p_resolveConflict() method should raise
are ConflictError and ValueError.  If it raises a different exception,
then it has failed.

In general, a unittest should not use a bare except clause, because
they almost always mask failures.  If the test really means to catch
all exceptions, it's best to make a comment to that effect.


git-svn-id: svn://svn.zope.org/repos/main@1402 62d5b8a3-27da-0310-9561-8e5933582275