Fix bug in description of FileStorage layout.
Simplify santiy checking of length fields in transaction metadata.
Replace the following test:
[1]    if ul > tl or dl > tl or el > tl or tl < (23+ul+dl+el):
with this one:
[2]    if tl < (23+ul+dl+el):

If any of the first three comparisons in [1] is true, then the final
test must also be true.  It's unlikely that any of the tests will be
true, so just do the final one.


git-svn-id: svn://svn.zope.org/repos/main@1409 62d5b8a3-27da-0310-9561-8e5933582275