Change DemoStorage to use BTrees.OOBTree instead of BTree.
Raise correct exceptions in abortVersion() and commitVersion().

Spread out some code that had the vertical white space removed.

Fix the versions() implementation so that it works if the optional max
argument is not passed by the caller.

Fix commitVersion() so that it works for committing from one version
to the other.  The previous version always committed to the
non-version data even if a destination version was specified.

Fix typos.


git-svn-id: svn://svn.zope.org/repos/main@1426 62d5b8a3-27da-0310-9561-8e5933582275