New transitions to provide for moving accepted and pending issues between confidential and public states:
 - restrict_pending, restrict_accepted: to confidential from public

 - unrestrict_pending, unrestrict_accepted: to public from
   confidential

The pending, pending_confidential, accepted, and accepted_confidential
states to make the appropriate transitions available.

There still is no way to have something confidential in completed
states - i'll wait until i see demand for it.

The workflow needs to be documented, in general.  It's probably now at
the state for it!  (Is there anyone out there wanting to get familiar
with DCWorkflow?  It may be productive to examine this one - it's
fairly substantial, and, i think, not misguided - and help out by
taking notes and providing them as the basis for documentation of this
workflow...)


git-svn-id: svn://svn.zope.org/repos/main@34138 62d5b8a3-27da-0310-9561-8e5933582275