Implemented issue editing.
In addition to invoking the issue .edit() method, we check for
toggling of security_related, and if seen, look for a workflow action
that restricts or unrestricts accordingly.  It makes sense for the
user interface to do this for the user, while it doesn't for the
application object business logic, so this is probably the right
place.  (It's also general enough so new restrict/unrestrict
transitions will be used, if/when added to the workflow.)

Return to the transcript, with a report of the changes.


git-svn-id: svn://svn.zope.org/repos/main@34142 62d5b8a3-27da-0310-9561-8e5933582275