Small change to recently added 'request' transition to make it robust in the absence of the 'requested' state variable.
This was prompted by a backwards-compatability problem for (already
valuable) issues in the collector development collector on
new.zope.org.  It may be justified for other reasons, but i realize
that my reasoning in my response to the problem report is wrong - any
issue created after the change will have the variable present when
checked.

Oh well - i'll have to consider tomorrow finding a way to add the
state variable to all the pre-existing issues (but that's something
DCWorkflow does not make easy, and may not be worth the trouble,
sigh).

The issue (and flawed analysis) is at
http://new.zope.org/Members/klm/ColDev/31


git-svn-id: svn://svn.zope.org/repos/main@34224 62d5b8a3-27da-0310-9561-8e5933582275