Enlarged the displayed title size (a lot!), and the description size (a bit).
Reorganized the browse-entry arrangement, i think for the better
(taking some cues from a custom view tres was using).  I eliminated
the number of uploads from the report, grouped the "from" and activity
dates on a line, and then grouped the status and all traits on the
next line - plus the supporters, when assigned, since there's room
there, and not in the previous line.


git-svn-id: svn://svn.zope.org/repos/main@31773 62d5b8a3-27da-0310-9561-8e5933582275