Use the new web text format, instead of stx.
.CookedBody(): For expediency, we apply format_webtext() to the
transcript's cooked body - which the processing happens on every
view.  What we *ought* to do is put the processing in the cooking of
the body - in a WebTextDocument derived from Document.  I actually
have such a beast, but it's sloppy, and won't take care of existing
issues.  When i have time to migrate them i'll take this step.


git-svn-id: svn://svn.zope.org/repos/main@31789 62d5b8a3-27da-0310-9561-8e5933582275