initTimeStamp(): Some cleanup of the initializer.  First be sure to decref the revision string, added to the module dict with the keys "__version__".  This fixes a small leaks detected by Insure.
Also, instead of returning from the init should the
PyString_FromString() of "TimeStamp.error fail, we simply take
precautions by XDECREF'ing it.  Then at the end of the function, we do
a PyErr_Occurred() check and throw a fatal error if true.  This makes
this module's init function more in line with other init functions in
this package.


git-svn-id: svn://svn.zope.org/repos/main@1481 62d5b8a3-27da-0310-9561-8e5933582275