initcPersistence(): Some cleanup of the initializer.  First be sure to decref the revision string, added to the module dict with the keys "__version__".  Do the same with the PyCObject set to the dict's "CAPI" key.  This fixes small leaks detected by Insure.
Also, clean up the way the global TimeStamp static is handled on
creation failure.  We need to decref `s' whether or not TimeStamp was
properly created or not.  Also, at the end of the function, we do a
PyErr_Occurred() check and throw a fatal error if true.  This makes
this module's init function more in line with other init functions in
this package.


git-svn-id: svn://svn.zope.org/repos/main@1482 62d5b8a3-27da-0310-9561-8e5933582275