Added readstats and writestats features which allow for saves and reuse of profile stats between runs of the requestprofiler.
Added urlfocus mode.  Urlfocus mode presents a summary of activity
within a period of time before and after the invocation of a particular
URL as it was recorded in the big M log.  This can be useful when trying
to track down a problem which you believe is related to some series
of URL invocations as opposed to a single URL invocation.

Fixed bug in timed mode which effected average stats.


git-svn-id: svn://svn.zope.org/repos/main@4609 62d5b8a3-27da-0310-9561-8e5933582275