(Possibly) correct use of Python memory APIs.
Fix SF bug #516768 reported by Dave Wallace.

Replace use of PyMem_DEL() with PyObject_Del() on object dealloc
functions.  The use of PyMem_DEL() is incorrect for object
deallocation, because it only ever calls the low-level free().  If a
custom allocator like pymalloc is used, it needs to be called to free
the memory.


git-svn-id: svn://svn.zope.org/repos/main@1722 62d5b8a3-27da-0310-9561-8e5933582275