checkTransactionalUndoAfterPack(): This is a test which should provoke the underlying bug in transactionalUndo() on a standby storage.  If our hypothesis is correct, the bug is in FileStorage, and is caused by encoding the file position in the `id' field of the undoLog information.  Note that Full just encodes the tid, but this is a problem for FileStorage (we have a strategy for fixing this).
NOTE: checking this in on the trunk is mildly antisocial because it
introduces failures in the test suite.  I'm doing it this way because
1) it really isn't worth a branch, even a short lived one; 2) Jeremy
may have time to look at it before tomorrow; 3) I don't trust my
machine not to lock up and make this mod unavailable. :(

Either way, this /will/ get fixed by tomorrow.


git-svn-id: svn://svn.zope.org/repos/main@1738 62d5b8a3-27da-0310-9561-8e5933582275