Improve error reporting when child exits or is killed by a signal.
Report explicitly what happened and include either the exit status or
the signal number.

XXX Not sure what to do if the process is just stopped.  The previous
version of the code treated it as an error.  This version doesn't
change that but does log a message saying its stopped.

XXX Not sure how to report whether a core was created.  The old code
did this implicitly if the platform/C library supported it.


git-svn-id: svn://svn.zope.org/repos/main@450 62d5b8a3-27da-0310-9561-8e5933582275