Make OkapiIndex the default index.
ZCTextIndex has grown a new argument with a default value that can be
used to specify an Index class to use.  The default is OkapiIndex.Index.

There is a little kludge to make the test succeed.
testZCTestIndex.IndexTests uses the Index.Index tests instead of
OkapiIndex.Index.  Tim will probably fix this.


git-svn-id: svn://svn.zope.org/repos/main@22177 62d5b8a3-27da-0310-9561-8e5933582275