It's really not safe for unit test modules to change sys.path.  Doing so makes the full suite brittle.  Cleaned out all the places where unit tests change sys.path.  (Use the PYTHONPATH environment variable instead.)
Also brought SearchIndex tests up to date, which I didn't really need to do
since SearchIndex is deprecated, but I did it as part of figuring out the
failed Interface tests, so I might as well check in my work. ;-)  One of the
test modules in SearchIndex forgets to define a global named "Dummy",
actually, so all the tests were failing before this checkin anyway.


git-svn-id: svn://svn.zope.org/repos/main@1889 62d5b8a3-27da-0310-9561-8e5933582275