Moved AbsoluteURL into a TraversalViews package, to cover both AbsoluteURL and also ObjectName.
ObjectName is sort of like a minimal version of AbsoluteURL in that it
gets you the name the current object was traversed via.
It can be used as a view, as @@object_name or by getting an adapter
for IObjectName, as in getAdapter(obj, IObjectName)().

Also, I added a __call__ method to IAbsoluteURL, as it was needed for
security declarations.


git-svn-id: svn://svn.zope.org/repos/main@5918 62d5b8a3-27da-0310-9561-8e5933582275