I hooked up Sidei's stylesheets and icon. Cleaned up th eHTML and CSS a bit too.
There was also a buglet in the TAL interpreter, where the message id was
sometimes None. This was due to a wrong usage of i18n:attributes, where
someone did:

<input name="Delete"
       i18n:attributes="name string:delete_button">

whereby the right form is:

<input name="Delete"
       tal:attributes="name string:delete_button"
       i18n:attributes="name">

After I removed the wrong syntax no None was sent to translate anymore, but
just to make sure I left the code inside.

I think that an exception should be raised, if i18n:attributes is used
incorrectly like this, but that is really Fred and Barry's department.


git-svn-id: svn://svn.zope.org/repos/main@5939 62d5b8a3-27da-0310-9561-8e5933582275